Extract post-order-handler body into service
Pull the order handler logic into a separate "Service" class. This class
preserves most of the existing behavior with some slight renames. It throws
named exceptions that are mapped to HTTP status codes in the handler.

Future iterations will properly inject this dependency. Some base classes will
fall out of this that can be used across the order types.