Add the OrderDispatcher
Create a generic "OrderDispatcher" that can be used to take a generic order / query and fire it off to the corresponding service. This lets the dispatcher be the primary entrypoint for create / listing orders so the handlers don't have to deal with the specifics of the interactions.

In the future, this will house a RelayOrderService and a CrossChainOrderService.